WIP: dynamically updating configuration
